Arm: Nouvelle vitesse et accélération shiftout des pid Ajout des valeurs de blocage des bras Maj des pid Set d'actions pour réséter le blocage comm_pc.c: Ajout de demande de log dans la commande générique + set des pids cvra_cs.c: Set des pids + régulation complete blocking_detection_manager.c: gestion du blocage sur 32 bits sans la fonction abs(16 bits) strat.c: fonction pour envoyer les logs d'un axe à la demande

+++ modules/blocking_detection_manager/blocking_detection_manager.h
@@ -44,14 +44,14 @@ struct blocking_detection {
 	struct cs *cs;
 	uint16_t cpt_thres;
 	uint16_t cpt;
-	uint16_t err_thres;
+	uint32_t err_thres;
 };
 
 /** init module, give the cs as parameter */
 void bd_init(struct blocking_detection *bd, struct cs *cs);
 
 
-void bd_set_thresholds(struct blocking_detection *bd, uint16_t err_thres, uint16_t cpt_thres);
+void bd_set_thresholds(struct blocking_detection *bd, uint32_t err_thres, uint16_t cpt_thres);
 
 /** reset the blocking */
 void bd_reset(struct blocking_detection *bd);
